Properly represent nodes in the NDB layer.

+++ src/Pst/Internal/Ndb/NbtEntry.cs
@@ -4,14 +4,14 @@ namespace Pst.Internal.Ndb
 {
     internal class NbtEntry : BTreeNode
     {
-        internal ulong Nid { get; set; }
+        internal uint Nid { get; set; }
         internal ulong DataBid { get; set; }
         internal ulong SubnodeBid { get; set; }
         internal uint ParentNid { get; set; }
 
         internal override void Create(byte[] data)
         {
-            Nid = BitConverter.ToUInt64(data, 0);
+            Nid = BitConverter.ToUInt32(data, 0);
             DataBid = BitConverter.ToUInt64(data, 8);
             SubnodeBid = BitConverter.ToUInt64(data, 16);
             ParentNid = BitConverter.ToUInt32(data, 24);
